Accept invalid From: header in email
This stops Cyrus from ruining everything all by itself; Your MTA is
the point where the validity of headers should be checked. For more
information on such and so forth, see (amongst others):

    http://www.postfix.org/header_checks.5.html

If one chooses to allow slightly invalid headers in the MTA, then
Cyrus should not raise problems.